feat: add useLogin hook
